+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.1.7](https://github.com/userfrosting/sprinkle-account/compare/5.1.6...5.1.7)
+- Fixes old session file not being deleted from the filesystem on logout
+
 ## [5.1.6](https://github.com/userfrosting/sprinkle-account/compare/5.1.5...5.1.6)
 - Add PHP 8.4 support & tests